requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
transformers>=4.35.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
Pillow>=9.5.0
scipy>=1.11.0
torchmetrics>=1.0.0
grad-cam>=1.4.8
netcal>=1.3.5
